requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
opencv-python>=4.8
scikit-image>=0.22
pytesseract>=0.3.10
PyMuPDF>=1.23
pdfplumber>=0.10
python-dotenv>=1.0
Pillow>=10.0
requests>=2.31
joblib>=1.3
imbalanced-learn>=0.11
kagglehub>=0.2.0
